requirements.txt
torch==2.0.0
torchvision==0.15.1
numpy==1.24.2
pillow==9.5.0
opencv-python==4.7.0.72
scipy==1.10.1
matplotlib==3.2.2
librosa==0.7.2
progressbar==2.5
pytorch-msssim==1.0.0
